symlink

+++ symlink.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 # ~/.local/bin
-if [ -d ~/.local/bin ]; then
+if ! [ -d ~/.local/bin ]; then
     mkdir -p ~/.local/bin
 fi
 
@@ -15,6 +15,7 @@ fi
 # bash
 if ! grep -q "# default distro ~/.bashrc above" ~/.bashrc; then
     cat >>~/.bashrc <<EOF
+
 # default distro ~/.bashrc above
 if [ -f "$HOME/.dotfiles/.bashrc" ]; then
     . "$HOME/.dotfiles/.bashrc"
@@ -41,8 +42,12 @@ else
     if command -v git &>/dev/null; then
         git clone https://github.com/tmux-plugins/tpm ~/.config/tmux/plugins/tpm
     fi
+fi
 
-    # tmux-sessionizer
+# tmux-sessionizer
+if [ -f ~/.local/bin/tmux-sessionizer.sh ]; then
+    echo "[info]: ~/.local/bin/tmux-sessionizer.sh already exists, skipping symlink creation."
+else
     ln -s ~/.dotfiles/bin/tmux-sessionizer.sh ~/.local/bin/.
 fi
 
@@ -76,3 +81,63 @@ if [ -f ~/.hidden ]; then
 else
     ln -s ~/.dotfiles/.hidden ~/.
 fi
+
+# #!/usr/bin/env bash
+#
+# # ~/.local/bin
+# if ! [ -d ~/.local/bin ]; then
+#     mkdir -p ~/.local/bin
+# fi
+#
+# # wezterm
+# mkdir -p ~/.config/wezterm
+# ln -s ~/.dotfiles/wezterm/wezterm.lua ~/.config/wezterm/.
+# ln -s ~/.dotfiles/wezterm/wezterm-notmux.lua ~/.config/wezterm/.
+# ln -s ~/.dotfiles/wezterm/utils.lua ~/.config/wezterm/.
+# ln -s ~/.dotfiles/wezterm/.style.lua ~/.config/wezterm/.
+#
+# # bash
+# if ! grep -q "# default .bashrc above" ~/.bashrc; then
+#     cat >>~/.bashrc <<EOF
+#
+# # default .bashrc above
+# if [ -f "$HOME/.dotfiles/.bashrc" ]; then
+#     . "$HOME/.dotfiles/.bashrc"
+# fi
+# EOF
+# fi
+#
+# # fish
+# mkdir -p ~/.config/fish
+# ln -s ~/.dotfiles/fish/config.fish ~/.config/fish/.
+# ln -s ~/.dotfiles/fish/alias.fish ~/.config/fish/.
+#
+# # tmux
+# if [ -d ~/.config/tmux ]; then
+#     echo "[info]: ~/.config/tmux already exists, skipping symlink creation."
+# else
+#     mkdir -p ~/.config/tmux
+#     ln -s ~/.dotfiles/tmux/tmux.conf ~/.config/tmux/.
+#
+#     if command -v git &>/dev/null; then
+#         git clone https://github.com/tmux-plugins/tpm ~/.config/tmux/plugins/tpm
+#     fi
+# fi
+#
+# # tmux-sessionizer
+# ln -s ~/.dotfiles/bin/tmux-sessionizer.sh ~/.local/bin/.
+#
+# # nvim
+# ln -s ~/.dotfiles/nvim ~/.config/.
+#
+# # bat
+# ln -s ~/.dotfiles/bat ~/.config/.
+# if command -v bat &>/dev/null; then
+#     /usr/bin/bat cache --build
+# fi
+#
+# # eza
+# ln -s ~/.dotfiles/eza ~/.config/.
+#
+# # hidden
+# ln -s ~/.dotfiles/.hidden ~/.
